--- pos/src/main/java/com/hexabyteit/api/pos/PosApplication.java
@@ -1,13 +1,0 @@
-package com.hexabyteit.api.pos;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class PosApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(PosApplication.class, args);
-	}
-
-}
