--- pos/src/main/java/com/hexabyteit/api/pos/model/constant/CompanyConstant.java
@@ -1,9 +1,0 @@
-package com.hexabyteit.api.pos.model.constant;
-
-public class CompanyConstant {
-
-    private CompanyConstant(){
-    }
-
-    public static final String TABLE_ORDER = "companies";
-}
