--- pos/src/main/java/com/hexabyteit/api/pos/model/modelservice/ResponseListCompany.java
@@ -1,18 +1,0 @@
-package com.hexabyteit.api.pos.model.modelservice;
-
-import com.hexabyteit.api.pos.model.Company;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-import java.util.List;
-
-@Builder
-@NoArgsConstructor
-@AllArgsConstructor
-@Data
-public class ResponseListCompany {
-
-    private List<Company> companies;
-}
