--- pos/src/main/java/com/hexabyteit/api/pos/repository/CompanyRepository.java
@@ -1,10 +1,0 @@
-package com.hexabyteit.api.pos.repository;
-
-import com.hexabyteit.api.pos.model.Company;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface CompanyRepository extends JpaRepository<Company, Long> {
-
-}
