--- pos/src/main/resources/application.properties
@@ -1,17 +1,0 @@
-##spring.datasource.url = jdbc:mysql://${RDS_HOSTNAME}:${RDS_PORT}/${RDS_DB_NAME}
-
-
-##spring.datasource.username=${RDS_USERNAME}
-##spring.datasource.password=${RDS_PASSWORD}
-
-spring.datasource.url=jdbc:postgresql://localhost:5432/POS
-// Cuenta (el valor predeterminado es postgres)
-spring.datasource.username=postgres
-// contraseña
-spring.datasource.password=postgres
-spring.datasource.driverClassName=org.postgresql.Driver
-spring.jpa.properties.hibernate.dialect=org.hibernate.dialect.PostgreSQLDialect
-
-spring.jpa.show-sql= true
-
-server.port=5000
