--- pos/src/test/java/com/hexabyteit/api/pos/PosApplicationTests.java
@@ -1,13 +1,0 @@
-package com.hexabyteit.api.pos;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class PosApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
