+++ pos_gradle/src/main/java/co/com/hexabyteit/pos/PosApplication.java
@@ -1,0 +1,13 @@
+package co.com.hexabyteit.pos;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class PosApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(PosApplication.class, args);
+	}
+
+}
