+++ pos_gradle/src/main/java/co/com/hexabyteit/pos/model/constant/CompanyConstant.java
@@ -1,0 +1,9 @@
+package co.com.hexabyteit.pos.model.constant;
+
+public class CompanyConstant {
+
+    private CompanyConstant(){
+    }
+
+    public static final String TABLE_ORDER = "companies";
+}
