+++ pos_gradle/src/main/java/co/com/hexabyteit/pos/repository/CompanyRepository.java
@@ -1,0 +1,7 @@
+package co.com.hexabyteit.pos.repository;
+
+import co.com.hexabyteit.pos.model.Company;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface CompanyRepository extends JpaRepository<Company, Long> {
+}
