+++ pos_gradle/src/main/resources/application.properties
@@ -1,0 +1,11 @@
+spring.datasource.url=jdbc:postgresql://localhost:5432/POS
+#Cuenta (el valor predeterminado es postgres)
+spring.datasource.username=postgres
+#contraseña
+spring.datasource.password=postgres
+spring.datasource.driverClassName=org.postgresql.Driver
+spring.jpa.properties.hibernate.dialect=org.hibernate.dialect.PostgreSQLDialect
+
+spring.jpa.show-sql= true
+
+server.port=5000
