+++ pos_gradle/src/test/java/co/com/hexabyteit/pos/PosApplicationTests.java
@@ -1,0 +1,13 @@
+package co.com.hexabyteit.pos;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class PosApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
